Add changes from the past few weeks
I've been working on the Metronome continuously to learn about Qt, QML
and such. I was not diligent and did not commit changes as I proceeded.
Hence this large commit.

Summary of changes:
- Many new component - used mainly for visualizations
  - Diode - two rectangles capable of color transition
  - FilledSlider (based on slider from ITU exercise)
  - TempoIndicators - basically just two text fields
  - PlayButton - play/pause button of metronome

- New metronome visualizations
  - MetronomeAnalogClassic
  - MetronomeSimple - currently WIP - will be finished in a separate PR
  - MetronomeAnalogModern - just a template, maybe will be skipped

Visualizations don't work correctly if executed separately. They need to
be given some properties and have signal handlers attached -> MVC.

- New "page"
  - MetronomePage

MetronomePage is the main area where the metronome visualizations will
be used. This should be the part where models and controllers are
applied.

- Several svgs used in the visualizations