Change file structure, add crude stackview, save metronome state
I changed the file structure because the number of qml files started
to grow considerably. Also their purpose was often different.

The stackview is currently showing one of two objects from
objectmodel. Ideally, this will be connected to some form of settings.
Also, there is a lot of duplication that needs to be taken out to KIS.

There is a new component MetronomeVisualization that serves as an
interface to the different forms of visualization. It is a simple
item with several properties. Visualizations are now based on it.

MetronomePage now saves the last tempo and tempoName. Again, this is
just temporary. Ideally, there will be a store of some kind where
strings, representing a tempo, will be mapped to some numerical
values. This store will probably be implemented in C++ as LUT in the
class exercise.