Make metronome visualizations scalable & sync slider positions
This required a lot of changes in how all the visualizations are done.
I tried to make use of property bindings as much as possible. The
outcome is not perfect but at least now the visualizations should be
able to fit a screen.

A new signal syncSliders has been added to MetronomeVisualization
because when scalling the sliders need to be kept in their correct
position. As a side effect of this, there is a way to centrally update
all sliders in a visualization.