Partially fix saving of last chosen tempo
While separating the visualization, saving of properties broke. This is
partially returning the behaviour. Tempo is saved on exit and properly
assigned to all existing visualizations but it is not shared between
visualizations themselves. For a change to take effect, the app has to
be restarted.

I'm not sure how to achieve the desired effect.